+++ inc/pro/bootstrap.php
@@ -5,5 +5,6 @@ require_once( HELIUM_THEME_INC . 'pro/customize-misc.php' );
 require_once( HELIUM_THEME_INC . 'pro/customize-social-media-contact.php' );
 require_once( HELIUM_THEME_INC . 'pro/customize-masonry.php' );
 require_once( HELIUM_THEME_INC . 'pro/customize-header.php' );
+require_once( HELIUM_THEME_INC . 'pro/customize-mobile.php' );
 require_once( HELIUM_THEME_INC . 'pro/admin-gradients.php' );
 require_once( HELIUM_THEME_INC . 'pro/functions-display.php' );